orders in admin

+++ app/Http/Controllers/Admin/OrdersController.php
@@ -8,11 +8,35 @@ use Illuminate\Http\Request;
 
 class OrdersController extends BaseController
 {
-    public function pending() {
+    public function newOrders() {
         $data = [
-            'title' => 'Заказы в ожидании',
+            'title' => 'Новые заказы',
         ];
-        $data['items'] = Order::getPendingOrders();
+        $data['items'] = Order::getOrdersWithStatus(Order::STATUS_NEW);
+        return view('admin.pages.orders.main', $data);
+    }
+
+    public function pendingOrders() {
+        $data = [
+            'title' => 'Невыполненные заказы',
+        ];
+        $data['items'] = Order::getOrdersWithStatus(Order::STATUS_PENDING);
+        return view('admin.pages.orders.main', $data);
+    }
+
+    public function doneOrders() {
+        $data = [
+            'title' => 'Выполненные заказы',
+        ];
+        $data['items'] = Order::getOrdersWithStatus(Order::STATUS_DONE);
+        return view('admin.pages.orders.main', $data);
+    }
+
+    public function declinedOrders() {
+        $data = [
+            'title' => 'Откланенные заказы',
+        ];
+        $data['items'] = Order::getOrdersWithStatus(Order::STATUS_NEW);
         return view('admin.pages.orders.main', $data);
     }
 
@@ -20,13 +44,48 @@ class OrdersController extends BaseController
         $data = [];
         $data['item'] = Order::getItem($id);
         $data['title'] = 'Заказ N'.$data['item']->id;
+        $data['process'] = Order::PROCESS;
         return view('admin.pages.orders.view', $data);
     }
 
     public function delete(Request $request) {
         $model = Order::getItem($request->input('id'));
+        if ($model->status != Order::STATUS_NEW && $model->status != Order::STATUS_DECLINED) return redirect()->back();
         Order::deleteItem($model);
         Notify::success('Заказ удален.');
         return redirect()->route('admin.orders.'.$model->status_type);
+    }
+
+    public function respond(Request $request, $id) {
+        $order = Order::getItem($id);
+        if ($order->status!==Order::STATUS_NEW) return redirect()->back();
+        if($request->input('status') == Order::STATUS_PENDING) {
+            $newStatus = Order::STATUS_PENDING;
+            $message = 'Заказ принят';
+        }
+        else {
+            $newStatus = Order::STATUS_DECLINED;
+            $message = 'Заяказ откланен';
+        }
+        $order->status = $newStatus;
+        $order->save();
+        Notify::success($message);
+        return redirect()->back();
+    }
+
+    public function changeProcess(Request $request, $id) {
+        $order = Order::getItem($id);
+        if ($order->status!= Order::STATUS_PENDING) return redirect()->back();
+        $process = (int) $request->input('process');
+        $paid = (int) $request->has('paid');
+        if ($process<0 || $process>3) return redirect()->back();
+        $order->process = $process;
+        $order->paid = $paid;
+        if ($process == 3 && $paid) {
+            $order->status = Order::STATUS_DONE;
+        }
+        $order->save();
+        Notify::get('changes_saved');
+        return redirect()->back();
     }
 }
